about-us page ui changed
