+++ .vscode/settings.json
@@ -12,5 +12,8 @@
     ],
     "[javascriptreact]": {
       "editor.defaultFormatter": "vscode.typescript-language-features"
+    },
+    "[javascript]": {
+      "editor.defaultFormatter": "vscode.typescript-language-features"
     }
   }
