+++ src/App.jsx
@@ -5,12 +5,12 @@ import PageOfflineGame from './pages/offline-game/PageOfflineGame'
 import PageUserHome from './pages/user-home/PageUserHome'
 import pages from './constants/pages'
 import PagePlayWithFriend from './pages/play-with-friend/PagePlayWithFriend'
-import PageAboutUs from './pages/about-us/PageAboutUs'
 import { ToastContextProvider } from './contexts/ToastContext'
 import { AlertContextProvider } from './contexts/AlertContext'
 import PageExtras from './pages/extras/PageExtras'
 import PageHowToPlay from './pages/how-to-play/PageHowToPlay'
 import { BackDropContextProvider } from './contexts/BackDropContext'
+import PageAboutUs from './pages/extras/about-us/PageAboutUs'
 
 function App () {
   return (
